Remove old listing webpage content
This content has now been moved to the `gh-pages` branch of the
`feedstocks` repo. Also the `feedstocks` webpage listing is being
generated by GitHub using the `gh-pages` branch. Given these changes,
there is no further need for the webpage content to live in the `master`
branch. Hence this clears this unneeded content out.