Remove `mkdocstrings-python-legacy`
This was created due to a temporary commit used to try and fix the
CircleCI registration issues encounted in staged-recipes. However the
commit no longer exists as the repo was removed after adding that
commit. In any event leaving this in here atm causes issues cloning the
repo, so remove it. Once the feedstock is converted this will be readded
anyways.